Remove the serde Serializer
