+++ meilisearch-core/src/serde/mod.rs
@@ -13,14 +13,12 @@ mod convert_to_string;
 mod deserializer;
 mod extract_document_id;
 mod indexer;
-mod serializer;
 
 pub use self::convert_to_number::ConvertToNumber;
 pub use self::convert_to_string::ConvertToString;
 pub use self::deserializer::{Deserializer, DeserializerError};
 pub use self::extract_document_id::{compute_document_id, extract_document_id, value_to_string};
 pub use self::indexer::Indexer;
-pub use self::serializer::{serialize_value, serialize_value_with_id, Serializer};
 
 use std::{error::Error, fmt};
